fix: limit minimum font size
On some platforms, setting an invalid 'guifont' option will cause the
option to be reset, preventing this plugin from further changing the
font size.

A font option of `h0` is considered invalid.

This prevents the font size from being decreased below 0, while still
letting it be minimized to 1.

Some front-ends (e.g. neovide) allow floating point font sizes, but I
don't feel it's worth the effort to allow the font size to be decreased
to an arbitrarily small amount below 1.